Tech_know_dictionary/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
itsdangerous==1.1.0
Flask-SQLAlchemy==2.3.2
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
nltk==3.5
regex==2020.9.27
tqdm==4.50.0
Werkzeug==1.0.1
WTForms==2.3.3
Flask-Bootstrap==3.3.7.1
gunicorn==20.0.4
pyspellchecker==0.5.5
bcrypt==3.1.4
blinker==1.4
certifi==2016.2.28
cffi==1.11.5
Flask-Bcrypt==0.7.1
Pillow==5.3.0
pycparser==2.18
six==1.11.0
SQLAlchemy==1.2.7